+++ blockchain/api/schemas.py
@@ -5,7 +5,8 @@ from typing import List, Optional, Union
 from fastapi import Path, Query
 from pydantic import AnyUrl, BaseModel, Field, PositiveInt, UUID4, constr
 
-from blockchain.impl import ConsentAction, ConsentType
+from blockchain.impl import ChangeHistoryStatus, ConsentAction, ConsentType, DataType
+from blockchain.typedefs import Mapping
 
 
 VersionString = constr(regex="^[0-9]+.[0-9]+.[0-9]+$")
@@ -65,10 +66,46 @@ class RegisterNodesResponse(BaseModel):
     total: int = Field(description="Amount of nodes that can participate in consensus resolution.")
 
 
+class SubSystemBase(BaseModel):
+    data_type: DataType = Field(description="Category of the data which consents are applied onto.")
+    data_description: Optional[str] = Field(description="Details providing more context about the consented data.")
+    data_source: Optional[str] = Field(description="Provenance of the data (e.g.: URL, services, etc.).")
+    data_provider: Optional[str] = Field(description="Data provider that generates the consents requirements.")
+    metadata: Optional[Union[Mapping, str]] = Field(
+        description="Additional metadata to be associated with the consents to better describe them."
+    )
+
+
+class SubSystemDataSubmitted(BaseModel):
+    media_type: Optional[str] = Field(
+        description="Explicit IANA data content format (defaults to plain/text or octets based on data if omitted)."
+    )
+    data: Union[bytes, str] = Field(
+        description="Literal data that contents apply to for validation (not stored in blockchain)."
+    )
+
+
+class SubSystemRequestBody(SubSystemBase, SubSystemDataSubmitted):
+    class Config:
+        extra = "forbid"  # make sure to forbid 'data_hash' for example
+
+
+class SubSystemDataResolved(BaseModel):
+    media_type: str = Field(description="Specified or resolved IANA data content format.")
+    data_hash: str = Field(description="Encrypted data representation of validation of associated consents.")
+
+
+class SubSystemResponseBody(SubSystemBase, SubSystemDataResolved):
+    pass
+
+
 class ConsentRequestBody(BaseModel):
     action: ConsentAction
     consent: bool
     expire: Optional[datetime] = None
+    subsystems: Optional[List[SubSystemRequestBody]] = Field(
+        description="Subsystem definitions and metadata related to data with applied consents."
+    )
 
 
 class ConsentSchema(BaseModel):
@@ -82,6 +119,9 @@ class ConsentSchema(BaseModel):
         "Revert back automatically to revoked consent when datetime is reached. "
         "Consents never expire if unspecified, unless explicitly updated with revoked consent."
     ))
+    subsystems: List[SubSystemResponseBody] = Field(
+        description="Subsystem definitions and metadata related to data with applied consents."
+    )
 
 
 class TransactionSchema(BaseModel):
@@ -124,12 +164,25 @@ class ChainConsentsDetailedResponse(BaseModel):
     length: PositiveInt
 
 
+class Change(BaseModel):
+    class Config:
+        extra = "forbid"
+
+    status: ChangeHistoryStatus = Field(description="Modification type that introduced consents change in the history.")
+    detail: str = Field(description="Description of the consent change in the history.")
+    action: Optional[ConsentAction] = Field(description="Applicable action that caused consent change in the history.")
+    consent: Optional[bool] = Field(description="Allowed or revoked consent action that cause change in the history.")
+    created: Optional[datetime] = Field(description="Creation datetime that caused consent change in the history.")
+    expired: Optional[datetime] = Field(description="Expiration datetime that caused consent change in the history.")
+
+
 class ListConsentsResponse(BaseModel):
     message: str = Field(description="Result of consents represented in the blockchain.")
     updated: datetime = Field(description="Last moment the chain was updated by consensus resolution.")
     outdated: bool = Field(description="Status indicating if a consent change was detected across the node network.")
     verified: bool = Field(description="Status indicating if latest consents were validated against other nodes.")
-    changes: List[str] = Field(description="Summary of all consent changes history applied over the chain.")
+    summary: List[str] = Field(description="Summary of all consent changes history applied over the chain.")
+    changes: List[Change] = Field(description="Detailed consent change history applied over the chain.")
     consents: List[ConsentSchema] = Field(description="Latest resolved consents.")