+++ blockchain/ui/views.py
@@ -10,10 +10,11 @@ from fastapi_mako import FastAPIMako  # noqa
 from pydantic import UUID4
 
 import blockchain
-from blockchain import AnyUUID, JSON, __meta__
+from blockchain import __meta__
 from blockchain.api import schemas
 from blockchain.api.chain import get_chain, get_chain_links, view_consents
 from blockchain.impl import Blockchain
+from blockchain.typedefs import AnyUUID, JSON
 from blockchain.utils import get_links
 
 
@@ -23,10 +24,51 @@ MAKO = FastAPIMako()
 
 
 def add_metadata(request: Request, data: Dict[str, Any]) -> Dict[str, Any]:
+    data["node_shortcuts"] = []
     data["node_id"] = request.app.node.id
     data["node_url"] = request.app.node.url
     data["version"] = __meta__["version"]
+    # generic CSS to applied on detected fields using 'get_styled_value' (see 'utils.mako')
+    # right side should be one of many CSS classes defined in 'styles.css'
+    data["styles"] = {
+        "action": "enum code nowrap",
+        "type": "enum code",
+        "consent": "bool",
+        "created": "date wrap",
+        "expire": "date wrap",
+        "data_type": "enum",
+        "data_provider": "",
+        "data_hash": "hash",
+        "bool": "bool",
+        "date": "date",
+        "datetime": "datetime",
+        "code": "code",
+        "enum": "enum",
+        "hash": "hash",
+        "uuid": "uuid",
+        "undefined": "undefined",
+        "wrap": "wrap",
+        "nowrap": "nowrap",
+        "lowercase": "lowercase",
+        "warning": "warning",
+        "label": "label",
+        "plain": "",
+    }
+    data["ui_shortcuts"] = get_ui_node_links(request)
     return data
+
+
+def get_ui_node_links(request):
+    """
+    Obtains generic UI links combined with specific UI endpoints for the current node.
+    """
+    links = get_links(request, VIEWS, self=False)
+    for link in links:
+        if link["rel"] == "ui":
+            link["title"] = "Node Network"
+        else:
+            link["title"] = link["rel"].replace("_", " ").capitalize()
+    return links
 
 
 def get_chain_shortcuts(request: Request, chain_id: AnyUUID) -> List[schemas.Link]:
@@ -59,15 +101,12 @@ def get_chain_info(request: Request, chain: Blockchain, strip_shortcut: Optional
 @VIEWS.get(
     "/",
     tags=["UI"],
-    summary="Display shortcuts to other pages.",
+    summary="Display Node details, networks nodes and shortcuts to other pages.",
     response_class=HTMLResponse,
 )
-@MAKO.template("ui/templates/view_shortcuts.mako")
+@MAKO.template("ui/templates/view_node.mako")
 async def shortcut_navigate(request: Request):
-    links = get_links(request, VIEWS, self=False)
-    for link in links:
-        link["title"] = link["rel"].replace("_", " ").capitalize()
-    data = {"links": links, "nodes": request.app.nodes}
+    data = {"nodes": request.app.nodes}
     return add_metadata(request, data)
 
 
@@ -136,4 +175,22 @@ async def display_consents(request: Request, chain_id: UUID4):
     chain = get_chain(request.app, chain_id)
     data = await view_consents(request, chain.id)
     data.update(get_chain_info(request, chain, strip_shortcut="consents"))
+    data.update({
+        "consent_fields": {
+            "action": "Action",
+            "consent": "Consent",
+            "type": "Type",
+            "created": "Created",
+            "expire": "Expire"
+        },
+        "subsystem_fields": {
+            "data_type": "Type",
+            "data_description": "Description",
+            "data_source": "Source",
+            "data_provider": "Provider",
+            "data_hash": "Hash",
+            "media_type": "Media-Type",
+            "metadata": "Extra Metadata",
+        }
+    })
     return add_metadata(request, data)